more more error handling to own function
